Add more usage of streams
The codebase already has some usage of streams where applicable.

More streams is being used to satisfy the A-Streams tag.

This commit adds more usage of streams in the Storage class.

Its added in the save method where the list of tasks is being written to the file
as I think that's where its most applicable.